refactor: resolve circular dependency between checkout components
